dashboard/requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
dj-database-url==0.5.0
Django==2.2.10
expertai-nlapi==2.1.3
gunicorn==20.0.4
idna==2.10
multidict==5.1.0
pycodestyle==2.6.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.26.3
whitenoise==5.2.0
yarl==1.6.3
